+++ CHANGELOG.md
@@ -2,10 +2,8 @@
 
 TODO:
 
-- Add border side, inline, block
-- Use border shadow in components
-- Move their usage in components to resolveMixin
 - Add mixin property to all components
+- Add border side, inline, block
 
 Highlights: